changed some counters to floats just in cased they need to be used as such. reduced features to just 2 signals x 4 so 8 features total for testing. for some reason sum_sigma is not getting updated. need to re-examine everything individually.

+++ MIT_DLEG/inc/back_estimation_methods.h
@@ -20,7 +20,7 @@ struct back_estimator_s
     float min_stance_theta;
     float max_stance_theta;
     float prev_min_stance_theta;
-    int prev_passed_ds_z_thresh_samples;
+    float prev_passed_ds_z_thresh_samples;
     int passed_ds_z_thresh_samples;
     uint8_t passed_us_z_thresh;
     uint8_t passed_ds_z_thresh;
